fix(desktop): address v2-workspaces review feedback
- Sidebar column sort direction was inverted (default "desc" put non-sidebar
  items first); use a normal boolean comparator so descending means "in
  sidebar first".
- Row onKeyDown was firing when Enter/Space bubbled from focused action
  buttons; ignore keyboard events originating from descendants.
- Offline host cell regressed to a native `title` tooltip; wrap it in the
  shared `<Tooltip>` so the indicator is keyboard-accessible and styled
  consistently with the action buttons.
- Extract SortableHeader into its own folder per repo conventions
  (one component per file) and move shared sort types into a types module.
