+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspacesList/components/SortableHeader/index.ts
@@ -1,0 +1,1 @@
+export { SortableHeader } from "./SortableHeader";
