+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspacesList/types.ts
@@ -1,0 +1,2 @@
+export type SortField = "sidebar" | "name" | "host" | "branch" | "created";
+export type SortDirection = "asc" | "desc";
